feat: add detail book with formatting
