+++ package.json
@@ -41,6 +41,7 @@
     "cmdk": "^1.0.0",
     "date-fns": "^3.6.0",
     "dayjs": "^1.11.11",
+    "html-react-parser": "^5.1.10",
     "lucide-react": "^0.378.0",
     "next-themes": "^0.3.0",
     "postcss": "^8.4.38",
